Split or reduce consistent cases
