+++ python/oneflow/test/modules/test_consistent_einsum_tensor_contraction.py
@@ -22,7 +22,7 @@ import oneflow.unittest
 from oneflow.test_utils.automated_test_util import *
 
 
-@autotest(n=2, check_graph=False)
+@autotest(n=1, check_graph=False)
 def _test_einsum_tensor_contraction(test_case, placement, sbp):
     dim0 = random(1, 3) * 8
     dim1 = random(1, 3) * 8
