+++ python/oneflow/test/modules/test_consistent_tril_with_diag.py
@@ -15,32 +15,34 @@ limitations under the License.
 """
 import unittest
 
-import numpy as np
-
+from oneflow.test_utils.automated_test_util import *
 import oneflow as flow
 import oneflow.unittest
-from oneflow.test_utils.automated_test_util import *
 
 
-@autotest(n=1, check_graph=False)
-def _test_einsum_batch_matmul2(test_case, placement, sbp):
-    dim0 = random(1, 3) * 8
-    dim1 = random(1, 3) * 8
-    dim2 = random(1, 3) * 8
-    x = random_tensor(ndim=4, dim0=dim0, dim1=dim1, dim2=random(1, 3) * 8, dim3=dim2)
-    y = random_tensor(ndim=4, dim0=dim0, dim1=dim1, dim2=dim2, dim3=random(1, 3) * 8)
-    g_x = x.to_global(placement=placement, sbp=sbp)
-    g_y = y.to_global(placement=placement, sbp=sbp)
-    z = torch.einsum("b h i j, b h j d -> b h i d", g_x, g_y)
-    return z
+
+@autotest(n=2, check_graph=False)
+def _test_global_tril_with_diag(test_case, placement, sbp):
+    diagonal = random(-3, 3).to(int)
+    x = random_tensor(
+        ndim=4,
+        dim0=random(1, 5).to(int) * 8,
+        dim1=random(1, 5).to(int) * 8,
+        dim2=random(1, 5).to(int) * 8,
+        dim3=random(1, 5).to(int) * 8,
+    ).to_global(placement, sbp)
+    y = torch.tril(x, diagonal)
+    y = torch.exp(y)
+
+    return y
 
 
-class TestEinsumConsistent(flow.unittest.TestCase):
+class TestConsistentTril(flow.unittest.TestCase):
     @globaltest
-    def test_einsum_batch_matmul2(test_case):
+    def test_global_tril_with_diag(test_case):
         for placement in all_placement():
             for sbp in all_sbp(placement, max_dim=4):
-                _test_einsum_batch_matmul2(test_case, placement, sbp)
+                _test_global_tril_with_diag(test_case, placement, sbp)
 
 
 if __name__ == "__main__":
