+++ python/oneflow/test/modules/test_consistent_tril_without_diag copy.py
@@ -20,7 +20,7 @@ import oneflow as flow
 import oneflow.unittest
 
 
-@autotest(n=2, check_graph=False)
+@autotest(n=1, check_graph=False)
 def _test_global_tril_without_diag(test_case, placement, sbp):
     x = random_tensor(
         ndim=4,
@@ -35,35 +35,12 @@ def _test_global_tril_without_diag(test_case, placement, sbp):
     return y
 
 
-@autotest(n=2, check_graph=False)
-def _test_global_tril_with_diag(test_case, placement, sbp):
-    diagonal = random(-3, 3).to(int)
-    x = random_tensor(
-        ndim=4,
-        dim0=random(1, 5).to(int) * 8,
-        dim1=random(1, 5).to(int) * 8,
-        dim2=random(1, 5).to(int) * 8,
-        dim3=random(1, 5).to(int) * 8,
-    ).to_global(placement, sbp)
-    y = torch.tril(x, diagonal)
-    y = torch.exp(y)
-
-    return y
-
-
 class TestConsistentTril(flow.unittest.TestCase):
     @globaltest
     def test_global_tril_without_diag(test_case):
         for placement in all_placement():
             for sbp in all_sbp(placement, max_dim=4):
                 _test_global_tril_without_diag(test_case, placement, sbp)
-
-    @globaltest
-    def test_global_tril_with_diag(test_case):
-        for placement in all_placement():
-            for sbp in all_sbp(placement, max_dim=4):
-                _test_global_tril_with_diag(test_case, placement, sbp)
-
 
 if __name__ == "__main__":
     unittest.main()
